// workflow_fe/src/lib/api/gen/types/UserResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UserResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  email: string;
  /**
   * @type string
   */
  employeeCode: string | null;
  /**
   * @type string
   */
  employeeGrade: string | null;
  /**
   * @type string
   */
  designation: string | null;
  /**
   * @type string
   */
  departmentId: string | null;
  /**
   * @type string
   */
  managerId: string | null;
  /**
   * @type boolean
   */
  isActive: boolean;
  /**
   * @type string, date-time
   */
  lastLoginAt: string | null;
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};